requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
python-nmap
pytest
httpx
python-json-logger
